Generic interface

Krylov.jl provides a generic interface for solving linear systems using (block) Krylov methods. The interface is designed to be common for all methods and contains three routines krylov_workspace, krylov_solve and krylov_solve!. They allow to build Krylov workspaces and call both in-place and out-of-place variants of the solvers with a unified API.

Krylov.krylov_workspace — Function

krylov_workspace(Val{method}(), args...; kwargs...)

Generic function that dispatches to the appropriate workspace constructor for each subtype of KrylovSolver and BlockKrylovSolver. The first argument Val{method}(), where method is a symbol (such as :cg, :gmres or :block_minres), specifies the (block) Krylov method for which a workspace is desired. The returned workspace can later be used by krylov_solve! to execute the (block) Krylov method in-place.

Krylov.krylov_solve! — Function

krylov_solve!(solver, args...; kwargs...)

Generic function that dispatches to the appropriate in-place (block) Krylov method based on the type of solver. The argument solver must be a subtype of KrylovSolver or BlockKrylovSolver (such as CgSolver, GmresSolver or BlockMinresSolver).

Krylov.krylov_solve — Function

krylov_solve(Val{method}(), args...; kwargs...)

Generic function that dispatches to the appropriate out-of-place (block) Krylov method specified by symbol method (such as :cg, :gmres or :block_minres).

In-place solvers update the workspace, from which solutions and statistics can be retrieved. The following functions are available for post-solve analysis:

Krylov.results — Function

results(solver)

Return a tuple containing the solution(s) and the statistics associated with the solver. Allows retrieving the output arguments of an out-of-place method from the in-place method.

For example, instead of x, stats = cg(A, b), you can use:

    solver = CgSolver(A, b)
    cg!(solver, A, b)
    x, stats = results(solver)

Krylov.solution — Function

solution(solver)

Return the solution(s) stored in the solver. Optionally you can specify which solution you want to recover, solution(solver, 1) returns x and solution(solver, 2) returns y.

Krylov.statistics — Function

statistics(solver)

Return the statistics stored in the solver.

Krylov.nsolution — Function

nsolution(solver)

Return the number of outputs of solution(solver).

Krylov.issolved — Function

issolved(solver)

Return a boolean that determines whether the Krylov method associated to solver succeeded.

Examples

using Krylov, Random

# Define a symmetric positive definite matrix A and a right-hand side vector b
n = 1000
A = sprandn(n, n, 0.005)
A = A * A' + I
b = randn(n)

# Out-of-place interface
for method in (:cg, :cr, :car)
    x, stats = krylov_solve(Val{method}(), A, b)
    r = b - A * x
    println("Residual norm for $(method): ", norm(r))
end

using Krylov, Random

# Define a square nonsymmetric matrix A and a right-hand side vector b
n = 100
A = sprand(n, n, 0.05) + I
b = rand(n)

# In-place interface
for method in (:bicgstab, :gmres)
    # Create a workspace for the Krylov method
    solver = krylov_workspace(Val(method), A, b)

    # Solve the system in-place
    krylov_solve!(solver, A, b)

    # Get the statistics
    stats = statistics(solver)

    # Retrieve the solution
    x = solution(solver)

    # Check if the solver converged
    solved = issolved(solver)
    println("Converged $method: ", solved)

    # Display the number of iterations
    niter = niterations(solver)
    println("Number of iterations for $method: ", niter)
end