Reference

Index

Krylov.kstdout

Krylov.AdjointStats

Krylov.BicgstabSolver

Krylov.BilqSolver

Krylov.BilqrSolver

Krylov.BlockGmresSolver

Krylov.BlockKrylovSolver

Krylov.BlockMinresSolver

Krylov.CarSolver

Krylov.CgLanczosShiftSolver

Krylov.CgLanczosSolver

Krylov.CgSolver

Krylov.CglsLanczosShiftSolver

Krylov.CglsSolver

Krylov.CgneSolver

Krylov.CgsSolver

Krylov.CrSolver

Krylov.CraigSolver

Krylov.CraigmrSolver

Krylov.CrlsSolver

Krylov.CrmrSolver

Krylov.DiomSolver

Krylov.DqgmresSolver

Krylov.FgmresSolver

Krylov.FloatOrComplex

Krylov.FomSolver

Krylov.GmresSolver

Krylov.GpmrSolver

Krylov.KrylovConstructor

Krylov.KrylovSolver

Krylov.KrylovStats

Krylov.LNLQStats

Krylov.LSLQStats

Krylov.LanczosShiftStats

Krylov.LanczosStats

Krylov.LnlqSolver

Krylov.LslqSolver

Krylov.LsmrSolver

Krylov.LsmrStats

Krylov.LsqrSolver

Krylov.MinaresSolver

Krylov.MinresQlpSolver

Krylov.MinresSolver

Krylov.QmrSolver

Krylov.SimpleStats

Krylov.SymmlqSolver

Krylov.SymmlqStats

Krylov.TricgSolver

Krylov.TrilqrSolver

Krylov.TrimrSolver

Krylov.UsymlqSolver

Krylov.UsymqrSolver

Base.show

Krylov.Aprod

Krylov.Atprod

Krylov.arnoldi

Krylov.arnoldi

Krylov.bicgstab

Krylov.bicgstab!

Krylov.bilq

Krylov.bilq!

Krylov.bilqr

Krylov.bilqr!

Krylov.block_gmres

Krylov.block_gmres!

Krylov.block_minres

Krylov.block_minres!

Krylov.car

Krylov.car!

Krylov.cg

Krylov.cg!

Krylov.cg_lanczos

Krylov.cg_lanczos!

Krylov.cg_lanczos_shift

Krylov.cg_lanczos_shift!

Krylov.cgls

Krylov.cgls!

Krylov.cgls_lanczos_shift

Krylov.cgls_lanczos_shift!

Krylov.cgne

Krylov.cgne!

Krylov.cgs

Krylov.cgs!

Krylov.cr

Krylov.cr!

Krylov.craig

Krylov.craig!

Krylov.craigmr

Krylov.craigmr!

Krylov.crls

Krylov.crls!

Krylov.crmr

Krylov.crmr!

Krylov.diom

Krylov.diom!

Krylov.dqgmres

Krylov.dqgmres!

Krylov.extract_parameters

Krylov.fgmres

Krylov.fgmres!

Krylov.fom

Krylov.fom!

Krylov.gmres

Krylov.gmres!

Krylov.golub_kahan

Krylov.golub_kahan

Krylov.gpmr

Krylov.gpmr!

Krylov.hermitian_lanczos

Krylov.hermitian_lanczos

Krylov.issolved

Krylov.kones

Krylov.ktypeof

Krylov.kzeros

Krylov.lnlq

Krylov.lnlq!

Krylov.lslq

Krylov.lslq!

Krylov.lsmr

Krylov.lsmr!

Krylov.lsqr

Krylov.lsqr!

Krylov.matrix_to_vector

Krylov.minares

Krylov.minares!

Krylov.minres

Krylov.minres!

Krylov.minres_qlp

Krylov.minres_qlp!

Krylov.montoison_orban

Krylov.montoison_orban

Krylov.niterations

Krylov.nonhermitian_lanczos

Krylov.nonhermitian_lanczos

Krylov.nsolution

Krylov.qmr

Krylov.qmr!

Krylov.results

Krylov.roots_quadratic

Krylov.saunders_simon_yip

Krylov.saunders_simon_yip

Krylov.solution

Krylov.solve!

Krylov.statistics

Krylov.sym_givens

Krylov.symmlq

Krylov.symmlq!

Krylov.to_boundary

Krylov.tricg

Krylov.tricg!

Krylov.trilqr

Krylov.trilqr!

Krylov.trimr

Krylov.trimr!

Krylov.usymlq

Krylov.usymlq!

Krylov.usymqr

Krylov.usymqr!

Krylov.vec2str

Krylov.vector_to_matrix

Krylov.FloatOrComplex — Type

FloatOrComplex{T}

Union type of T and Complex{T} where T is an AbstractFloat.

Krylov.niterations — Function

niterations(solver)

Return the number of iterations performed by the Krylov method associated to solver.

Krylov.Aprod — Function

Aprod(solver)

Return the number of operator-vector products with A performed by the Krylov method associated to solver.

Krylov.Atprod — Function

Atprod(solver)

Return the number of operator-vector products with A' performed by the Krylov method associated to solver.

Krylov.kstdout — Constant

Default I/O stream for all Krylov methods.

Krylov.extract_parameters — Function

arguments = extract_parameters(ex::Expr)

Extract the arguments of an expression that is keyword parameter tuple. Implementation suggested by Mitchell J. O'Sullivan (@mosullivan93).

Base.show — Function

show(io, solver; show_stats=true)

Statistics of solver are displayed if show_stats is set to true.