Detailed examples may be found here.

Krylov methods

Krylov.cg — Function

(x, stats) = cg(A, b::AbstractVector{T};
                M=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                itmax::Int=0, radius::T=zero(T), linesearch::Bool=false,
                verbose::Int=0) where T <: AbstractFloat

The conjugate gradient method to solve the symmetric linear system Ax=b.

The method does not abort if A is not definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite. M also indicates the weighted norm in which residuals are measured.

If itmax=0, the default number of iterations is set to 2 * n, with n = length(b).

Reference

M. R. Hestenes and E. Stiefel, Methods of conjugate gradients for solving linear systems, Journal of Research of the National Bureau of Standards, 49(6), pp. 409–436, 1952.

Krylov.cr — Function

(x, stats) = cr(A, b::AbstractVector{T};
                M=opEye(), atol::T=√eps(T), rtol::T=√eps(T), γ::T=√eps(T), itmax::Int=0,
                radius::T=zero(T), verbose::Int=0, linesearch::Bool=false) where T <: AbstractFloat

A truncated version of Stiefel’s Conjugate Residual method to solve the symmetric linear system Ax = b or the least-squares problem min ‖b - Ax‖. The matrix A must be positive semi-definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite. M also indicates the weighted norm in which residuals are measured.

In a linesearch context, 'linesearch' must be set to 'true'.

If itmax=0, the default number of iterations is set to 2 * n, with n = length(b).

References

M. R. Hestenes and E. Stiefel, Methods of conjugate gradients for solving linear systems, Journal of Research of the National Bureau of Standards, 49(6), pp. 409–436, 1952.

M-A. Dahito and D. Orban, The Conjugate Residual Method in Linesearch and Trust-Region Methods, SIAM Journal on Optimization, 29(3), pp. 1988–2025, 2019.

Krylov.symmlq — Function

(x, stats) = symmlq(A, b::AbstractVector{T};
                    M=opEye(), λ::T=zero(T), transfer_to_cg::Bool=true,
                    λest::T=zero(T), atol::T=√eps(T), rtol::T=√eps(T),
                    etol::T=√eps(T), window::Int=0, itmax::Int=0,
                    conlim::T=1/√eps(T), verbose::Int=0) where T <: AbstractFloat

Solve the shifted linear system

(A + λ I) x = b

using the SYMMLQ method, where λ is a shift parameter, and A is square and symmetric.

SYMMLQ produces monotonic errors ‖x*-x‖₂.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Reference

C. C. Paige and M. A. Saunders, Solution of Sparse Indefinite Systems of Linear Equations, SIAM Journal on Numerical Analysis, 12(4), pp. 617–629, 1975.

Krylov.cg_lanczos — Function

(x, stats) = cg_lanczos(A, b::AbstractVector{T};
                        M=opEye(), atol::T=√eps(T), rtol::T=√eps(T), itmax::Int=0,
                        check_curvature::Bool=false, verbose::Int=0) where T <: AbstractFloat

The Lanczos version of the conjugate gradient method to solve the symmetric linear system

Ax = b

The method does not abort if A is not definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

References

A. Frommer and P. Maass, Fast CG-Based Methods for Tikhonov-Phillips Regularization, SIAM Journal on Scientific Computing, 20(5), pp. 1831–1850, 1999.

C. C. Paige and M. A. Saunders, Solution of Sparse Indefinite Systems of Linear Equations, SIAM Journal on Numerical Analysis, 12(4), pp. 617–629, 1975.

Krylov.cg_lanczos_shift_seq — Function

(x, stats) = cg_lanczos_shift_seq(A, b::AbstractVector{T}, shifts::AbstractVector{T};
                                  M=opEye(), atol::T=√eps(T), rtol::T=√eps(T), itmax::Int=0,
                                  check_curvature::Bool=false, verbose::Int=0) where T <: AbstractFloat

The Lanczos version of the conjugate gradient method to solve a family of shifted systems

(A + αI) x = b  (α = α₁, ..., αₙ)

The method does not abort if A + αI is not definite.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Krylov.minres — Function

(x, stats) = minres(A, b::AbstractVector{T};
                    M=opEye(), λ::T=zero(T), atol::T=√eps(T)/100,
                    rtol::T=√eps(T)/100, etol::T=√eps(T),
                    window::Int=5, itmax::Int=0, conlim::T=1/√eps(T),
                    verbose::Int=0) where T <: AbstractFloat

Solve the shifted linear least-squares problem

minimize ‖b - (A + λ I)x‖₂²

or the shifted linear system

(A + λ I) x = b

using the MINRES method, where λ ≥ 0 is a shift parameter, where A is square and symmetric.

MINRES is formally equivalent to applying CR to Ax=b when A is positive definite, but is typically more stable and also applies to the case where A is indefinite.

MINRES produces monotonic residuals ‖r‖₂ and optimality residuals ‖Aᵀr‖₂.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite.

Reference

C. C. Paige and M. A. Saunders, Solution of Sparse Indefinite Systems of Linear Equations, SIAM Journal on Numerical Analysis, 12(4), pp. 617–629, 1975.

Krylov.minres_qlp — Function

(x, stats) = minres_qlp(A, b::AbstractVector{T};
                        M=opEye(), atol::T=√eps(T), rtol::T=√eps(T), λ::T=zero(T),
                        itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

MINRES-QLP is the only method based on the Lanczos process that returns the minimum-norm solution on singular inconsistent systems (A + λI)x = b, where λ is a shift parameter. It is significantly more complex but can be more reliable than MINRES when A is ill-conditioned.

A preconditioner M may be provided in the form of a linear operator and is assumed to be symmetric and positive definite. M also indicates the weighted norm in which residuals are measured.

References

S.-C. T. Choi, Iterative methods for singular linear equations and least-squares problems, Ph.D. thesis, ICME, Stanford University, 2006.

S.-C. T. Choi, C. C. Paige and M. A. Saunders, MINRES-QLP: A Krylov subspace method for indefinite or singular symmetric systems, SIAM Journal on Scientific Computing, Vol. 33(4), pp. 1810–1836, 2011.

S.-C. T. Choi and M. A. Saunders, Algorithm 937: MINRES-QLP for symmetric and Hermitian linear equations and least-squares problems, ACM Transactions on Mathematical Software, 40(2), pp. 1–12, 2014.

Krylov.diom — Function

(x, stats) = diom(A, b::AbstractVector{T};
                  M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T), itmax::Int=0,
                  memory::Int=20, pivoting::Bool=false, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system Ax = b using direct incomplete orthogonalization method.

DIOM is similar to CG with partial reorthogonalization.

An advantage of DIOM is that nonsymmetric or symmetric indefinite or both nonsymmetric and indefinite systems of linear equations can be handled by this single algorithm.

This implementation allows a left preconditioner M and a right preconditioner N.

Left preconditioning : M⁻¹Ax = M⁻¹b

Right preconditioning : AN⁻¹u = b with x = N⁻¹u

Split preconditioning : M⁻¹AN⁻¹u = M⁻¹b with x = N⁻¹u

Reference

Y. Saad, Practical use of some krylov subspace methods for solving indefinite and nonsymmetric linear systems, SIAM journal on scientific and statistical computing, 5(1), pp. 203–228, 1984.

Krylov.dqgmres — Function

(x, stats) = dqgmres(A, b::AbstractVector{T};
                     M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                     itmax::Int=0, memory::Int=20, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system Ax = b using DQGMRES method.

DQGMRES algorithm is based on the incomplete Arnoldi orthogonalization process and computes a sequence of approximate solutions with the quasi-minimal residual property.

This implementation allows a left preconditioner M and a right preconditioner N.

Left preconditioning : M⁻¹Ax = M⁻¹b

Right preconditioning : AN⁻¹u = b with x = N⁻¹u

Split preconditioning : M⁻¹AN⁻¹u = M⁻¹b with x = N⁻¹u

Reference

Y. Saad and K. Wu, DQGMRES: a quasi minimal residual algorithm based on incomplete orthogonalization, Numerical Linear Algebra with Applications, Vol. 3(4), pp. 329–343, 1996.

Krylov.usymlq — Function

(x, stats) = usymlq(A, b::AbstractVector{T}, c::AbstractVector{T};
                    atol::T=√eps(T), rtol::T=√eps(T), transfer_to_usymcg::Bool=true,
                    itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the linear system Ax = b using the USYMLQ method.

USYMLQ is based on a tridiagonalization process for unsymmetric matrices. The error norm ‖x - x*‖ monotonously decreases in USYMLQ. It's considered as a generalization of SYMMLQ.

It can also be applied to under-determined and over-determined problems. In all cases, problems must be consistent.

An option gives the possibility of transferring to the USYMCG point, when it exists. The transfer is based on the residual norm.

References

M. A. Saunders, H. D. Simon, and E. L. Yip, Two Conjugate-Gradient-Type Methods for Unsymmetric Linear Equations, SIAM Journal on Numerical Analysis, 25(4), pp. 927–940, 1988.

A. Buttari, D. Orban, D. Ruiz and D. Titley-Peloquin, A tridiagonalization method for symmetric saddle-point and quasi-definite systems, SIAM Journal on Scientific Computing, 41(5), pp. 409–432, 2019.

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.usymqr — Function

(x, stats) = usymqr(A, b::AbstractVector{T}, c::AbstractVector{T};
                    atol::T=√eps(T), rtol::T=√eps(T),
                    itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the linear system Ax = b using the USYMQR method.

USYMQR is based on a tridiagonalization process for unsymmetric matrices. The residual norm ‖b - Ax‖ monotonously decreases in USYMQR. It's considered as a generalization of MINRES.

It can also be applied to under-determined and over-determined problems. USYMQR finds the minimum-norm solution if problems are inconsistent.

References

M. A. Saunders, H. D. Simon, and E. L. Yip, Two Conjugate-Gradient-Type Methods for Unsymmetric Linear Equations, SIAM Journal on Numerical Analysis, 25(4), pp. 927–940, 1988.

A. Buttari, D. Orban, D. Ruiz and D. Titley-Peloquin, A tridiagonalization method for symmetric saddle-point and quasi-definite systems, SIAM Journal on Scientific Computing, 41(5), pp. 409–432, 2019.

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.tricg — Function

(x, y, stats) = tricg(A, b::AbstractVector{T}, c::AbstractVector{T};
                      M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                      spd::Bool=false, snd::Bool=false, flip::Bool=false,
                      τ::T=one(T), ν::T=-one(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

TriCG solves the symmetric linear system

[ τE    A ] [ x ] = [ b ]
[  Aᵀ  νF ] [ y ]   [ c ],

where τ and ν are real numbers, E = M⁻¹ ≻ 0 and F = N⁻¹ ≻ 0. TriCG could breakdown if τ = 0 or ν = 0. It's recommended to use TriMR in these cases.

By default, TriCG solves symmetric and quasi-definite linear systems with τ = 1 and ν = -1. If flip = true, TriCG solves another known variant of SQD systems where τ = -1 and ν = 1. If spd = true, τ = ν = 1 and the associated symmetric and positive definite linear system is solved. If snd = true, τ = ν = -1 and the associated symmetric and negative definite linear system is solved. τ and ν are also keyword arguments that can be directly modified for more specific problems.

TriCG is based on the preconditioned orthogonal tridiagonalization process and its relation with the preconditioned block-Lanczos process.

[ M   O ]
[ 0   N ]

indicates the weighted norm in which residuals are measured. It's the Euclidean norm when M and N are identity operators.

TriCG stops when itmax iterations are reached or when ‖rₖ‖ ≤ atol + ‖r₀‖ * rtol. atol is an absolute tolerance and rtol is a relative tolerance.

Additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every verbose iterations.

Reference

A. Montoison and D. Orban, TriCG and TriMR: Two Iterative Methods for Symmetric Quasi-Definite Systems, Cahier du GERAD G-2020-41, GERAD, Montréal, 2020.

Krylov.trimr — Function

(x, y, stats) = trimr(A, b::AbstractVector{T}, c::AbstractVector{T};
                      M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                      spd::Bool=false, snd::Bool=false, flip::Bool=false, sp::Bool=false,
                      τ::T=one(T), ν::T=-one(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

TriMR solves the symmetric linear system

[ τE    A ] [ x ] = [ b ]
[  Aᵀ  νF ] [ y ]   [ c ],

where τ and ν are real numbers, E = M⁻¹ ≻ 0 and F = N⁻¹ ≻ 0. TriMR handles saddle-point systems (τ = 0 or ν = 0) and adjoint systems (τ = 0 and ν = 0) without any risk of breakdown.

By default, TriMR solves symmetric and quasi-definite linear systems with τ = 1 and ν = -1. If flip = true, TriMR solves another known variant of SQD systems where τ = -1 and ν = 1. If spd = true, τ = ν = 1 and the associated symmetric and positive definite linear system is solved. If snd = true, τ = ν = -1 and the associated symmetric and negative definite linear system is solved. If sp = true, τ = 1, ν = 0 and the associated saddle-point linear system is solved. τ and ν are also keyword arguments that can be directly modified for more specific problems.

TriMR is based on the preconditioned orthogonal tridiagonalization process and its relation with the preconditioned block-Lanczos process.

[ M   O ]
[ 0   N ]

indicates the weighted norm in which residuals are measured. It's the Euclidean norm when M and N are identity operators.

TriMR stops when itmax iterations are reached or when ‖rₖ‖ ≤ atol + ‖r₀‖ * rtol. atol is an absolute tolerance and rtol is a relative tolerance.

Additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every verbose iterations.

Reference

A. Montoison and D. Orban, TriCG and TriMR: Two Iterative Methods for Symmetric Quasi-Definite Systems, Cahier du GERAD G-2020-41, GERAD, Montréal, 2020.

Krylov.trilqr — Function

(x, t, stats) = trilqr(A, b::AbstractVector{T}, c::AbstractVector{T};
                       atol::T=√eps(T), rtol::T=√eps(T), transfer_to_usymcg::Bool=true,
                       itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Combine USYMLQ and USYMQR to solve adjoint systems.

[0  A] [t] = [b]
[Aᵀ 0] [x]   [c]

USYMLQ is used for solving primal system Ax = b. USYMQR is used for solving dual system Aᵀt = c.

An option gives the possibility of transferring from the USYMLQ point to the USYMCG point, when it exists. The transfer is based on the residual norm.

Reference

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.bilq — Function

(x, stats) = bilq(A, b::AbstractVector{T}; c::AbstractVector{T}=b,
                  atol::T=√eps(T), rtol::T=√eps(T), transfer_to_bicg::Bool=true,
                  itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the square linear system Ax = b using the BiLQ method.

BiLQ is based on the Lanczos biorthogonalization process. When A is symmetric and b = c, BiLQ is equivalent to SYMMLQ.

An option gives the possibility of transferring to the BiCG point, when it exists. The transfer is based on the residual norm.

Reference

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.cgs — Function

(x, stats) = cgs(A, b::AbstractVector{T}; c::AbstractVector{T}=b,
                 M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                 itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system Ax = b using conjugate gradient squared algorithm.

From "Iterative Methods for Sparse Linear Systems (Y. Saad)" :

«The method is based on a polynomial variant of the conjugate gradients algorithm. Although related to the so-called bi-conjugate gradients (BCG) algorithm, it does not involve adjoint matrix-vector multiplications, and the expected convergence rate is about twice that of the BCG algorithm.

The Conjugate Gradient Squared algorithm works quite well in many cases. However, one difficulty is that, since the polynomials are squared, rounding errors tend to be more damaging than in the standard BCG algorithm. In particular, very high variations of the residual vectors often cause the residual norms computed to become inaccurate.

TFQMR and BICGSTAB were developed to remedy this difficulty.»

This implementation allows a left preconditioner M and a right preconditioner N.

Reference

P. Sonneveld, CGS, A Fast Lanczos-Type Solver for Nonsymmetric Linear systems, SIAM Journal on Scientific and Statistical Computing, 10(1), pp. 36–52, 1989.

Krylov.bicgstab — Function

(x, stats) = bicgstab(A, b::AbstractVector{T}; c::AbstractVector{T}=b,
                      M=opEye(), N=opEye(), atol::T=√eps(T), rtol::T=√eps(T),
                      itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the square linear system Ax = b using the BICGSTAB method.

The Biconjugate Gradient Stabilized method is a variant of BiCG, like CGS, but using different updates for the Aᵀ-sequence in order to obtain smoother convergence than CGS.

If BICGSTAB stagnates, we recommend DQGMRES and BiLQ as alternative methods for unsymmetric square systems.

BICGSTAB stops when itmax iterations are reached or when ‖rₖ‖ ≤ atol + ‖b‖ * rtol. atol is an absolute tolerance and rtol is a relative tolerance.

Additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every verbose iterations.

This implementation allows a left preconditioner M and a right preconditioner N.

References

H. A. van der Vorst, Bi-CGSTAB: A fast and smoothly converging variant of Bi-CG for the solution of nonsymmetric linear systems, SIAM Journal on Scientific and Statistical Computing, 13(2), pp. 631–644, 1992.

G. L.G. Sleijpen and D. R. Fokkema, BiCGstab(ℓ) for linear equations involving unsymmetric matrices with complex spectrum, Electronic Transactions on Numerical Analysis, 1, pp. 11–32, 1993.

Krylov.qmr — Function

(x, stats) = qmr(A, b::AbstractVector{T}; c::AbstractVector{T}=b,
                 atol::T=√eps(T), rtol::T=√eps(T),
                 itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the square linear system Ax = b using the QMR method.

QMR is based on the Lanczos biorthogonalization process. When A is symmetric and b = c, QMR is equivalent to MINRES.

References

R. W. Freund and N. M. Nachtigal, QMR : a quasi-minimal residual method for non-Hermitian linear systems, Numerische mathematik, Vol. 60(1), pp. 315–339, 1991.

R. W. Freund and N. M. Nachtigal, An implementation of the QMR method based on coupled two-term recurrences, SIAM Journal on Scientific Computing, Vol. 15(2), pp. 313–337, 1994.

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.bilqr — Function

(x, t, stats) = bilqr(A, b::AbstractVector{T}, c::AbstractVector{T};
                      atol::T=√eps(T), rtol::T=√eps(T), transfer_to_bicg::Bool=true,
                      itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Combine BiLQ and QMR to solve adjoint systems.

[0  A] [t] = [b]
[Aᵀ 0] [x]   [c]

BiLQ is used for solving primal system Ax = b. QMR is used for solving dual system Aᵀt = c.

An option gives the possibility of transferring from the BiLQ point to the BiCG point, when it exists. The transfer is based on the residual norm.

Reference

A. Montoison and D. Orban, BiLQ: An Iterative Method for Nonsymmetric Linear Systems with a Quasi-Minimum Error Property, SIAM Journal on Matrix Analysis and Applications, 41(3), pp. 1145–1166, 2020.

Krylov.cgls — Function

(x, stats) = cgls(A, b::AbstractVector{T};
                  M=opEye(), λ::T=zero(T), atol::T=√eps(T), rtol::T=√eps(T),
                  radius::T=zero(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ ‖x‖₂²

using the Conjugate Gradient (CG) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CG to the normal equations

(AᵀA + λI) x = Aᵀb

but is more stable.

CGLS produces monotonic residuals ‖r‖₂ but not optimality residuals ‖Aᵀr‖₂. It is formally equivalent to LSQR, though can be slightly less accurate, but simpler to implement.

References

M. R. Hestenes and E. Stiefel. Methods of conjugate gradients for solving linear systems, Journal of Research of the National Bureau of Standards, 49(6), pp. 409–436, 1952.

A. Björck, T. Elfving and Z. Strakos, Stability of Conjugate Gradient and Lanczos Methods for Linear Least Squares Problems, SIAM Journal on Matrix Analysis and Applications, 19(3), pp. 720–736, 1998.

Krylov.crls — Function

(x, stats) = crls(A, b::AbstractVector{T};
                  M=opEye(), λ::T=zero(T), atol::T=√eps(T), rtol::T=√eps(T),
                  radius::T=zero(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the linear least-squares problem

minimize ‖b - Ax‖₂² + λ ‖x‖₂²

using the Conjugate Residuals (CR) method. This method is equivalent to applying MINRES to the normal equations

(AᵀA + λI) x = Aᵀb.

This implementation recurs the residual r := b - Ax.

CRLS produces monotonic residuals ‖r‖₂ and optimality residuals ‖Aᵀr‖₂. It is formally equivalent to LSMR, though can be substantially less accurate, but simpler to implement.

Reference

D. C.-L. Fong, Minimum-Residual Methods for Sparse, Least-Squares using Golubg-Kahan Bidiagonalization, Ph.D. Thesis, Stanford University, 2011.

Krylov.cgne — Function

(x, stats) = cgne(A, b::AbstractVector{T};
                  M=opEye(), λ::T=zero(T), atol::T=√eps(T), rtol::T=√eps(T),
                  itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system

Ax + √λs = b

using the Conjugate Gradient (CG) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CG to the normal equations of the second kind

(AAᵀ + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂ s.t. Ax = b.

When λ > 0, it solves the problem

min ‖(x,s)‖₂  s.t. Ax + √λs = b.

CGNE produces monotonic errors ‖x-x*‖₂ but not residuals ‖r‖₂. It is formally equivalent to CRAIG, though can be slightly less accurate, but simpler to implement. Only the x-part of the solution is returned.

A preconditioner M may be provided in the form of a linear operator.

References

J. E. Craig, The N-step iteration procedures, Journal of Mathematics and Physics, 34(1), pp. 64–73, 1955.

J. E. Craig, Iterations Procedures for Simultaneous Equations, Ph.D. Thesis, Department of Electrical Engineering, MIT, 1954.

Krylov.crmr — Function

(x, stats) = crmr(A, b::AbstractVector{T};
                  M=opEye(), λ::T=zero(T), atol::T=√eps(T),
                  rtol::T=√eps(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system

Ax + √λs = b

using the Conjugate Residual (CR) method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying CR to the normal equations of the second kind

(AAᵀ + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂  s.t.  x ∈ argmin ‖Ax - b‖₂.

When λ > 0, this method solves the problem

min ‖(x,s)‖₂  s.t. Ax + √λs = b.

CGMR produces monotonic residuals ‖r‖₂. It is formally equivalent to CRAIG-MR, though can be slightly less accurate, but simpler to implement. Only the x-part of the solution is returned.

A preconditioner M may be provided in the form of a linear operator.

References

D. Orban and M. Arioli, Iterative Solution of Symmetric Quasi-Definite Linear Systems, Volume 3 of Spotlights. SIAM, Philadelphia, PA, 2017.

D. Orban, The Projected Golub-Kahan Process for Constrained Linear Least-Squares Problems. Cahier du GERAD G-2014-15, 2014.

Krylov.lslq — Function

(x_lq, x_cg, err_lbnds, err_ubnds_lq, err_ubnds_cg, stats) =
    lslq(A, b::AbstractVector{T};
         M=opEye(), N=opEye(), sqd::Bool=false, λ::T=zero(T),
         atol::T=√eps(T), btol::T=√eps(T), etol::T=√eps(T),
         window::Int=5, utol::T=√eps(T), itmax::Int=0,
         σ::T=zero(T), conlim::T=1/√eps(T), verbose::Int=0) where T <: AbstractFloat

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSLQ method, where λ ≥ 0 is a regularization parameter. LSLQ is formally equivalent to applying SYMMLQ to the normal equations

(AᵀA + λ² I) x = Aᵀb

but is more stable.

Main features

the solution estimate is updated along orthogonal directions

the norm of the solution estimate ‖xᴸₖ‖₂ is increasing

the error ‖eₖ‖₂ := ‖xᴸₖ - x*‖₂ is decreasing

it is possible to transition cheaply from the LSLQ iterate to the LSQR iterate if there is an advantage (there always is in terms of error)

if A is rank deficient, identify the minimum least-squares solution

Input arguments

A::AbstractLinearOperator

b::Vector{Float64}

Optional arguments

M::AbstractLinearOperator=opEye(): a symmetric and positive definite dual preconditioner

N::AbstractLinearOperator=opEye(): a symmetric and positive definite primal preconditioner

sqd::Bool=false indicates whether or not we are solving a symmetric and quasi-definite augmented system

If sqd = true, we solve the symmetric and quasi-definite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ  -F ] [ x ] = [ 0 ],

where E and F are symmetric and positive definite. LSLQ is then equivalent to applying SYMMLQ to (AᵀE⁻¹A + F)y = AᵀE⁻¹b with r = E⁻¹(b - Ax). Preconditioners M = E⁻¹ ≻ 0 and N = F⁻¹ ≻ 0 may be provided in the form of linear operators.

If sqd is set to false (the default), we solve the symmetric and indefinite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ   0 ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the weighted norm in which x and Aᵀr should be measured. r can be recovered by computing E⁻¹(b - Ax).

λ::Float64=0.0 is a regularization parameter (see the problem statement above)

σ::Float64=0.0 is an underestimate of the smallest nonzero singular value of A–-setting σ too large will result in an error in the course of the iterations

atol::Float64=1.0e-8 is a stopping tolerance based on the residual

btol::Float64=1.0e-8 is a stopping tolerance used to detect zero-residual problems

etol::Float64=1.0e-8 is a stopping tolerance based on the lower bound on the error

window::Int=5 is the number of iterations used to accumulate a lower bound on the error

utol::Float64=1.0e-8 is a stopping tolerance based on the upper bound on the error

itmax::Int=0 is the maximum number of iterations (0 means no imposed limit)

conlim::Float64=1.0e+8 is the limit on the estimated condition number of A beyond which the solution will be abandoned

verbose::Int=0 determines verbosity.

Return values

lslq() returns the tuple (x_lq, x_cg, err_lbnds, err_ubnds_lq, err_ubnds_cg, stats) where

x_lq::Vector{Float64} is the LQ solution estimate

x_cg::Vector{Float64} is the CG solution estimate (i.e., the LSQR point)

err_lbnds::Vector{Float64} is a vector of lower bounds on the LQ error–-the vector is empty if window is set to zero

err_ubnds_lq::Vector{Float64} is a vector of upper bounds on the LQ error–-the vector is empty if σ == 0 is left at zero

err_ubnds_cg::Vector{Float64} is a vector of upper bounds on the CG error–-the vector is empty if σ == 0 is left at zero

stats::SimpleStats collects other statistics on the run.

Stopping conditions

The iterations stop as soon as one of the following conditions holds true:

the optimality residual is sufficiently small (stats.status = "found approximate minimum least-squares solution") in the sense that either

‖Aᵀr‖ / (‖A‖ ‖r‖) ≤ atol, or

1 + ‖Aᵀr‖ / (‖A‖ ‖r‖) ≤ 1

an approximate zero-residual solution has been found (stats.status = "found approximate zero-residual solution") in the sense that either

‖r‖ / ‖b‖ ≤ btol + atol ‖A‖ * ‖xᴸ‖ / ‖b‖, or

1 + ‖r‖ / ‖b‖ ≤ 1

the estimated condition number of A is too large in the sense that either

1/cond(A) ≤ 1/conlim (stats.status = "condition number exceeds tolerance"), or

1 + 1/cond(A) ≤ 1 (stats.status = "condition number seems too large for this machine")

the lower bound on the LQ forward error is less than etol * ‖xᴸ‖

the upper bound on the CG forward error is less than utol * ‖xᶜ‖

References

R. Estrin, D. Orban and M. A. Saunders, Euclidean-norm error bounds for SYMMLQ and CG, SIAM Journal on Matrix Analysis and Applications, 40(1), pp. 235–253, 2019.

R. Estrin, D. Orban and M. A. Saunders, LSLQ: An Iterative Method for Linear Least-Squares with an Error Minimization Property, SIAM Journal on Matrix Analysis and Applications, 40(1), pp. 254–275, 2019.

Krylov.lsqr — Function

(x, stats) = lsqr(A, b::AbstractVector{T};
                  M=opEye(), N=opEye(), sqd::Bool=false,
                  λ::T=zero(T), axtol::T=√eps(T), btol::T=√eps(T),
                  atol::T=zero(T), rtol::T=zero(T),
                  etol::T=√eps(T), window::Int=5,
                  itmax::Int=0, conlim::T=1/√eps(T),
                  radius::T=zero(T), verbose::Int=0) where T <: AbstractFloat

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSQR method, where λ ≥ 0 is a regularization parameter. LSQR is formally equivalent to applying CG to the normal equations

(AᵀA + λ² I) x = Aᵀb

(and therefore to CGLS) but is more stable.

LSQR produces monotonic residuals ‖r‖₂ but not optimality residuals ‖Aᵀr‖₂. It is formally equivalent to CGLS, though can be slightly more accurate.

Preconditioners M and N may be provided in the form of linear operators and are assumed to be symmetric and positive definite. If sqd is set to true, we solve the symmetric and quasi-definite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ  -F ] [ x ] = [ 0 ],

where E and F are symmetric and positive definite. LSQR is then equivalent to applying CG to (AᵀE⁻¹A + F)y = AᵀE⁻¹b with r = E⁻¹(b - Ax). Preconditioners M = E⁻¹ ≻ 0 and N = F⁻¹ ≻ 0 may be provided in the form of linear operators.

If sqd is set to false (the default), we solve the symmetric and indefinite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ   0 ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the weighted norm in which x and Aᵀr should be measured. r can be recovered by computing E⁻¹(b - Ax).

Reference

C. C. Paige and M. A. Saunders, LSQR: An Algorithm for Sparse Linear Equations and Sparse Least Squares, ACM Transactions on Mathematical Software, 8(1), pp. 43–71, 1982.

Krylov.lsmr — Function

(x, stats) = lsmr(A, b::AbstractVector{T};
                  M=opEye(), N=opEye(), sqd::Bool=false,
                  λ::T=zero(T), axtol::T=√eps(T), btol::T=√eps(T),
                  atol::T=zero(T), rtol::T=zero(T),
                  etol::T=√eps(T), window::Int=5,
                  itmax::Int=0, conlim::T=1/√eps(T),
                  radius::T=zero(T), verbose::Int=0) where T <: AbstractFloat

Solve the regularized linear least-squares problem

minimize ‖b - Ax‖₂² + λ² ‖x‖₂²

using the LSMR method, where λ ≥ 0 is a regularization parameter. LSQR is formally equivalent to applying MINRES to the normal equations

(AᵀA + λ² I) x = Aᵀb

(and therefore to CRLS) but is more stable.

LSMR produces monotonic residuals ‖r‖₂ and optimality residuals ‖Aᵀr‖₂. It is formally equivalent to CRLS, though can be substantially more accurate.

Preconditioners M and N may be provided in the form of linear operators and are assumed to be symmetric and positive definite. If sqd is set to true, we solve the symmetric and quasi-definite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ  -F ] [ x ] = [ 0 ],

where E and F are symmetric and positive definite. LSMR is then equivalent to applying MINRES to (AᵀE⁻¹A + F)y = AᵀE⁻¹b with r = E⁻¹(b - Ax). Preconditioners M = E⁻¹ ≻ 0 and N = F⁻¹ ≻ 0 may be provided in the form of linear operators.

If sqd is set to false (the default), we solve the symmetric and indefinite system

[ E    A ] [ r ]   [ b ]
[ Aᵀ   0 ] [ x ] = [ 0 ].

In this case, N can still be specified and indicates the weighted norm in which x and Aᵀr should be measured. r can be recovered by computing E⁻¹(b - Ax).

Reference

D. C.-L. Fong and M. A. Saunders, LSMR: An Iterative Algorithm for Sparse, Least Squares Problems, SIAM Journal on Scientific Computing, 33(5), pp. 2950–2971, 2011.

Krylov.craig — Function

(x, y, stats) = craig(A, b::AbstractVector{T};
                      M=opEye(), N=opEye(), sqd::Bool=false, λ::T=zero(T), atol::T=√eps(T),
                      btol::T=√eps(T), rtol::T=√eps(T), conlim::T=1/√eps(T), itmax::Int=0,
                      verbose::Int=0, transfer_to_lsqr::Bool=false) where T <: AbstractFloat

Find the least-norm solution of the consistent linear system

Ax + λs = b

using the Golub-Kahan implementation of Craig's method, where λ ≥ 0 is a regularization parameter. This method is equivalent to CGNE but is more stable.

For a system in the form Ax = b, Craig's method is equivalent to applying CG to AAᵀy = b and recovering x = Aᵀy. Note that y are the Lagrange multipliers of the least-norm problem

minimize ‖x‖  s.t.  Ax = b.

Preconditioners M⁻¹ and N⁻¹ may be provided in the form of linear operators and are assumed to be symmetric and positive definite. If sqd = true, CRAIG solves the symmetric and quasi-definite system

[ -N   Aᵀ ] [ x ]   [ 0 ]
[  A   M  ] [ y ] = [ b ],

which is equivalent to applying CG to (AN⁻¹Aᵀ + M)y = b with Nx = Aᵀy.

If sqd = false, CRAIG solves the symmetric and indefinite system

[ -N   Aᵀ ] [ x ]   [ 0 ]
[  A   0  ] [ y ] = [ b ].

In this case, M⁻¹ can still be specified and indicates the weighted norm in which residuals are measured.

In this implementation, both the x and y-parts of the solution are returned.

References

C. C. Paige and M. A. Saunders, LSQR: An Algorithm for Sparse Linear Equations and Sparse Least Squares, ACM Transactions on Mathematical Software, 8(1), pp. 43–71, 1982.

M. A. Saunders, Solutions of Sparse Rectangular Systems Using LSQR and CRAIG, BIT Numerical Mathematics, 35(4), pp. 588–604, 1995.

Krylov.craigmr — Function

(x, y, stats) = craigmr(A, b::AbstractVector{T};
                        M=opEye(), N=opEye(), λ::T=zero(T), atol::T=√eps(T),
                        rtol::T=√eps(T), itmax::Int=0, verbose::Int=0) where T <: AbstractFloat

Solve the consistent linear system

Ax + √λs = b

using the CRAIG-MR method, where λ ≥ 0 is a regularization parameter. This method is equivalent to applying the Conjugate Residuals method to the normal equations of the second kind

(AAᵀ + λI) y = b

but is more stable. When λ = 0, this method solves the minimum-norm problem

min ‖x‖₂  s.t.  x ∈ argmin ‖Ax - b‖₂.

When λ > 0, this method solves the problem

min ‖(x,s)‖₂  s.t. Ax + √λs = b.

Preconditioners M⁻¹ and N⁻¹ may be provided in the form of linear operators and are assumed to be symmetric and positive definite. Afterward CRAIGMR solves the symmetric and quasi-definite system

[ -N   Aᵀ ] [ x ]   [ 0 ]
[  A   M  ] [ y ] = [ b ],

which is equivalent to applying MINRES to (M + AN⁻¹Aᵀ)y = b.

CRAIGMR produces monotonic residuals ‖r‖₂. It is formally equivalent to CRMR, though can be slightly more accurate, and intricate to implement. Both the x- and y-parts of the solution are returned.

References

D. Orban and M. Arioli. Iterative Solution of Symmetric Quasi-Definite Linear Systems, Volume 3 of Spotlights. SIAM, Philadelphia, PA, 2017.

D. Orban, The Projected Golub-Kahan Process for Constrained, Linear Least-Squares Problems. Cahier du GERAD G-2014-15, 2014.