Stats Types

Krylov.KrylovStats — Type

Abstract type for statistics returned by a solver

Krylov.SimpleStats — Type

Type for statistics returned by the majority of Krylov solvers, the attributes are:

solved

inconsistent

residuals

Aresiduals

status

Krylov.LanczosStats — Type

Type for statistics returned by Lanczos solvers CG-LANCZOS and CG-LANCZOS-SHIFT-SEQ, the attributes are:

solved

residuals

flagged

Anorm

Acond

status

Krylov.SymmlqStats — Type

Type for statistics returned by SYMMLQ, the attributes are:

solved

residuals

residualscg

errors

errorscg

Anorm

Acond

status

Krylov.AdjointStats — Type

Type for statistics returned by adjoint systems solvers BiLQR and TriLQR, the attributes are:

solved_primal

solved_dual

residuals_primal

residuals_dual

status

Solver Types

Krylov.KrylovSolver — Type

Abstract type for using Krylov solvers in-place

Krylov.MinresSolver — Type

Type for storing the vectors required by the in-place version of MINRES.

The outer constructor

solver = MinresSolver(A, b; window :: Int=5)

may be used in order to create these vectors.

Krylov.CgSolver — Type

Type for storing the vectors required by the in-place version of CG.

The outer constructor

solver = CgSolver(A, b)

may be used in order to create these vectors.

Krylov.CrSolver — Type

Type for storing the vectors required by the in-place version of CR.

The outer constructor

solver = CrSolver(A, b)

may be used in order to create these vectors.

Krylov.SymmlqSolver — Type

Type for storing the vectors required by the in-place version of SYMMLQ.

The outer constructor

solver = SymmlqSolver(A, b)

may be used in order to create these vectors.

Krylov.CgLanczosSolver — Type

Type for storing the vectors required by the in-place version of CG-LANCZOS.

The outer constructor

solver = CgLanczosSolver(A, b)

may be used in order to create these vectors.

Krylov.CgLanczosShiftSolver — Type

Type for storing the vectors required by the in-place version of CG-LANCZOS-SHIFT-SEQ.

The outer constructor

solver = CgLanczosShiftSolver(A, b)

may be used in order to create these vectors.

Krylov.MinresQlpSolver — Type

Type for storing the vectors required by the in-place version of MINRES-QLP.

The outer constructor

solver = MinresQlpSolver(A, b)

may be used in order to create these vectors.

Krylov.DqgmresSolver — Type

Type for storing the vectors required by the in-place version of DQGMRES.

The outer constructor

solver = DqgmresSolver(A, b; memory :: Integer=20)

may be used in order to create these vectors.

Krylov.DiomSolver — Type

Type for storing the vectors required by the in-place version of DIOM.

The outer constructor

solver = DiomSolver(A, b; memory :: Integer=20)

may be used in order to create these vectors.

Krylov.UsymlqSolver — Type

Type for storing the vectors required by the in-place version of USYMLQ.

The outer constructor

solver = UsymlqSolver(A, b)

may be used in order to create these vectors.

Krylov.UsymqrSolver — Type

Type for storing the vectors required by the in-place version of USYMQR.

The outer constructor

solver = UsymqrSolver(A, b)

may be used in order to create these vectors.

Krylov.TricgSolver — Type

Type for storing the vectors required by the in-place version of TRICG.

The outer constructor

solver = TricgSolver(A, b)

may be used in order to create these vectors.

Krylov.TrimrSolver — Type

Type for storing the vectors required by the in-place version of TRIMR.

The outer constructor

solver = TrimrSolver(A, b)

may be used in order to create these vectors.

Krylov.TrilqrSolver — Type

Type for storing the vectors required by the in-place version of TRILQR.

The outer constructor

solver = TrilqrSolver(A, b)

may be used in order to create these vectors.

Krylov.CgsSolver — Type

Type for storing the vectors required by the in-place version of CGS.

The outer constructor

solver = CgsSolver(A, b)

may be used in order to create these vectors.

Krylov.BicgstabSolver — Type

Type for storing the vectors required by the in-place version of BICGSTAB.

The outer constructor

solver = BicgstabSolver(A, b)

may be used in order to create these vectors.

Krylov.BilqSolver — Type

Type for storing the vectors required by the in-place version of BILQ.

The outer constructor

solver = BilqSolver(A, b)

may be used in order to create these vectors.

Krylov.QmrSolver — Type

Type for storing the vectors required by the in-place version of QMR.

The outer constructor

solver = QmrSolver(A, b)

may be used in order to create these vectors.

Krylov.BilqrSolver — Type

Type for storing the vectors required by the in-place version of BILQR.

The outer constructor

solver = BilqrSolver(A, b)

may be used in order to create these vectors.

Krylov.CglsSolver — Type

Type for storing the vectors required by the in-place version of CGLS.

The outer constructor

solver = CglsSolver(A, b)

may be used in order to create these vectors.

Krylov.CrlsSolver — Type

Type for storing the vectors required by the in-place version of CRLS.

The outer constructor

solver = CrlsSolver(A, b)

may be used in order to create these vectors.

Krylov.CgneSolver — Type

Type for storing the vectors required by the in-place version of CGNE.

The outer constructor

solver = CgneSolver(A, b)

may be used in order to create these vectors.

Krylov.CrmrSolver — Type

Type for storing the vectors required by the in-place version of CRMR.

The outer constructor

solver = CrmrSolver(A, b)

may be used in order to create these vectors.

Krylov.LslqSolver — Type

Type for storing the vectors required by the in-place version of LSLQ.

The outer constructor

solver = LslqSolver(A, b)

may be used in order to create these vectors.

Krylov.LsqrSolver — Type

Type for storing the vectors required by the in-place version of LSQR.

The outer constructor

solver = LsqrSolver(A, b)

may be used in order to create these vectors.

Krylov.LsmrSolver — Type

Type for storing the vectors required by the in-place version of LSMR.

The outer constructor

solver = LsmrSolver(A, b)

may be used in order to create these vectors.

Krylov.LnlqSolver — Type

Type for storing the vectors required by the in-place version of LNLQ.

The outer constructor

solver = LnlqSolver(A, b)

may be used in order to create these vectors.

Krylov.CraigSolver — Type

Type for storing the vectors required by the in-place version of CRAIG.

The outer constructor

solver = CraigSolver(A, b)

may be used in order to create these vectors.

Krylov.CraigmrSolver — Type

Type for storing the vectors required by the in-place version of CRAIGMR.

The outer constructor

solver = CraigmrSolver(A, b)

may be used in order to create these vectors.

Utilities

Krylov.roots_quadratic — Function

roots = roots_quadratic(q₂, q₁, q₀; nitref)

Find the real roots of the quadratic

q(x) = q₂ x² + q₁ x + q₀,

where q₂, q₁ and q₀ are real. Care is taken to avoid numerical cancellation. Optionally, nitref steps of iterative refinement may be performed to improve accuracy. By default, nitref=1.

Krylov.sym_givens — Function

(c, s, ρ) = sym_givens(a, b)

Numerically stable symmetric Givens reflection. Given a and b reals, return (c, s, ρ) such that

[ c  s ] [ a ] = [ ρ ]
[ s -c ] [ b ] = [ 0 ].

Numerically stable symmetric Givens reflection. Given a and b complexes, return (c, s, ρ) with c real and (s, ρ) complexes such that

[ c   s ] [ a ] = [ ρ ]
[ s̅  -c ] [ b ] = [ 0 ].

Krylov.to_boundary — Function

roots = to_boundary(x, d, radius; flip, xNorm2, dNorm2)

Given a trust-region radius radius, a vector x lying inside the trust-region and a direction d, return σ1 and σ2 such that

‖x + σi d‖ = radius, i = 1, 2

in the Euclidean norm. If known, ‖x‖² may be supplied in xNorm2.

If flip is set to true, σ1 and σ2 are computed such that

‖x - σi d‖ = radius, i = 1, 2.

Krylov.vec2str — Function

s = vec2str(x; ndisp)

Display an array in the form

[ -3.0e-01 -5.1e-01  1.9e-01 ... -2.3e-01 -4.4e-01  2.4e-01 ]

with (ndisp - 1)/2 elements on each side.

Krylov.ktypeof — Function

S = ktypeof(v)

Return a dense storage type S based on the type of v.

Krylov.kzeros — Function

v = kzeros(S, n)

Create an AbstractVector of storage type S of length n only composed of zero.

Krylov.kones — Function

v = kones(S, n)

Create an AbstractVector of storage type S of length n only composed of one.