Reference

Index

Krylov.AdjointStats

Krylov.BicgstabSolver

Krylov.BilqSolver

Krylov.BilqrSolver

Krylov.CgLanczosShiftSolver

Krylov.CgLanczosSolver

Krylov.CgSolver

Krylov.CglsSolver

Krylov.CgneSolver

Krylov.CgsSolver

Krylov.CrSolver

Krylov.CraigSolver

Krylov.CraigmrSolver

Krylov.CrlsSolver

Krylov.CrmrSolver

Krylov.DiomSolver

Krylov.DqgmresSolver

Krylov.FloatOrComplex

Krylov.FomSolver

Krylov.GmresSolver

Krylov.GpmrSolver

Krylov.KrylovSolver

Krylov.KrylovStats

Krylov.LNLQStats

Krylov.LSLQStats

Krylov.LanczosShiftStats

Krylov.LanczosStats

Krylov.LnlqSolver

Krylov.LslqSolver

Krylov.LsmrSolver

Krylov.LsqrSolver

Krylov.MinresQlpSolver

Krylov.MinresSolver

Krylov.QmrSolver

Krylov.SimpleStats

Krylov.SymmlqSolver

Krylov.SymmlqStats

Krylov.TricgSolver

Krylov.TrilqrSolver

Krylov.TrimrSolver

Krylov.UsymlqSolver

Krylov.UsymqrSolver

Base.show

Krylov.Aprod

Krylov.Atprod

Krylov.bicgstab

Krylov.bicgstab!

Krylov.bilq

Krylov.bilq!

Krylov.bilqr

Krylov.bilqr!

Krylov.cg

Krylov.cg!

Krylov.cg_lanczos

Krylov.cg_lanczos!

Krylov.cgls

Krylov.cgls!

Krylov.cgne

Krylov.cgne!

Krylov.cgs

Krylov.cgs!

Krylov.cr

Krylov.cr!

Krylov.craig

Krylov.craig!

Krylov.craigmr

Krylov.craigmr!

Krylov.crls

Krylov.crls!

Krylov.crmr

Krylov.crmr!

Krylov.diom

Krylov.diom!

Krylov.dqgmres

Krylov.dqgmres!

Krylov.fom

Krylov.fom!

Krylov.gmres

Krylov.gmres!

Krylov.gpmr

Krylov.gpmr!

Krylov.issolved

Krylov.kones

Krylov.ktypeof

Krylov.kzeros

Krylov.lnlq

Krylov.lnlq!

Krylov.lslq

Krylov.lslq!

Krylov.lsmr

Krylov.lsmr!

Krylov.lsqr

Krylov.lsqr!

Krylov.minres

Krylov.minres!

Krylov.minres_qlp

Krylov.minres_qlp!

Krylov.niterations

Krylov.nsolution

Krylov.qmr

Krylov.qmr!

Krylov.roots_quadratic

Krylov.solution

Krylov.solve!

Krylov.statistics

Krylov.sym_givens

Krylov.symmlq

Krylov.symmlq!

Krylov.to_boundary

Krylov.tricg

Krylov.tricg!

Krylov.trilqr

Krylov.trilqr!

Krylov.trimr

Krylov.trimr!

Krylov.usymlq

Krylov.usymlq!

Krylov.usymqr

Krylov.usymqr!

Krylov.vec2str

Krylov.FloatOrComplex — Type

FloatOrComplex{T}

Union type of T and Complex{T} where T is an AbstractFloat.

Krylov.niterations — Function

niterations(solver)

Return the number of iterations performed by the Krylov method associated to solver.

Krylov.Aprod — Function

Aprod(solver)

Return the number of operator-vector products with A performed by the Krylov method associated to solver.

Krylov.Atprod — Function

Atprod(solver)

Return the number of operator-vector products with A' performed by the Krylov method associated to solver.

Base.show — Function

show(io, solver; show_stats=true)

Statistics of solver are displayed if show_stats is set to true.