Reference

Index

Krylov.kstdout

Krylov.AdjointStats

Krylov.BicgstabWorkspace

Krylov.BilqWorkspace

Krylov.BilqrWorkspace

Krylov.BlockGmresWorkspace

Krylov.BlockKrylovWorkspace

Krylov.BlockMinresWorkspace

Krylov.CarWorkspace

Krylov.CgLanczosShiftWorkspace

Krylov.CgLanczosWorkspace

Krylov.CgWorkspace

Krylov.CglsLanczosShiftWorkspace

Krylov.CglsWorkspace

Krylov.CgneWorkspace

Krylov.CgsWorkspace

Krylov.CrWorkspace

Krylov.CraigWorkspace

Krylov.CraigmrWorkspace

Krylov.CrlsWorkspace

Krylov.CrmrWorkspace

Krylov.DiomWorkspace

Krylov.DqgmresWorkspace

Krylov.FgmresWorkspace

Krylov.FloatOrComplex

Krylov.FomWorkspace

Krylov.GmresWorkspace

Krylov.GpmrWorkspace

Krylov.KrylovConstructor

Krylov.KrylovStats

Krylov.KrylovWorkspace

Krylov.LNLQStats

Krylov.LSLQStats

Krylov.LanczosShiftStats

Krylov.LanczosStats

Krylov.LnlqWorkspace

Krylov.LslqWorkspace

Krylov.LsmrStats

Krylov.LsmrWorkspace

Krylov.LsqrWorkspace

Krylov.MinaresWorkspace

Krylov.MinresQlpWorkspace

Krylov.MinresWorkspace

Krylov.QmrWorkspace

Krylov.SimpleStats

Krylov.SymmlqStats

Krylov.SymmlqWorkspace

Krylov.TricgWorkspace

Krylov.TrilqrWorkspace

Krylov.TrimrWorkspace

Krylov.UsymlqWorkspace

Krylov.UsymlqrWorkspace

Krylov.UsymqrWorkspace

Base.show

Krylov.Aprod_count

Krylov.Atprod_count

Krylov.arnoldi

Krylov.arnoldi

Krylov.bicgstab

Krylov.bicgstab!

Krylov.bilq

Krylov.bilq!

Krylov.bilqr

Krylov.bilqr!

Krylov.block_gmres

Krylov.block_gmres!

Krylov.block_minres

Krylov.block_minres!

Krylov.car

Krylov.car!

Krylov.cg

Krylov.cg!

Krylov.cg_lanczos

Krylov.cg_lanczos!

Krylov.cg_lanczos_shift

Krylov.cg_lanczos_shift!

Krylov.cgls

Krylov.cgls!

Krylov.cgls_lanczos_shift

Krylov.cgls_lanczos_shift!

Krylov.cgne

Krylov.cgne!

Krylov.cgs

Krylov.cgs!

Krylov.cr

Krylov.cr!

Krylov.craig

Krylov.craig!

Krylov.craigmr

Krylov.craigmr!

Krylov.crls

Krylov.crls!

Krylov.crmr

Krylov.crmr!

Krylov.diom

Krylov.diom!

Krylov.dqgmres

Krylov.dqgmres!

Krylov.elapsed_allocation_time

Krylov.elapsed_time

Krylov.extract_parameters

Krylov.fgmres

Krylov.fgmres!

Krylov.fom

Krylov.fom!

Krylov.gmres

Krylov.gmres!

Krylov.golub_kahan

Krylov.golub_kahan

Krylov.gpmr

Krylov.gpmr!

Krylov.hermitian_lanczos

Krylov.hermitian_lanczos

Krylov.issolved

Krylov.iteration_count

Krylov.krylov_solve

Krylov.krylov_solve!

Krylov.krylov_workspace

Krylov.ktypeof

Krylov.lnlq

Krylov.lnlq!

Krylov.lslq

Krylov.lslq!

Krylov.lsmr

Krylov.lsmr!

Krylov.lsqr

Krylov.lsqr!

Krylov.matrix_to_vector

Krylov.minares

Krylov.minares!

Krylov.minres

Krylov.minres!

Krylov.minres_qlp

Krylov.minres_qlp!

Krylov.montoison_orban

Krylov.montoison_orban

Krylov.nonhermitian_lanczos

Krylov.nonhermitian_lanczos

Krylov.qmr

Krylov.qmr!

Krylov.results

Krylov.roots_quadratic

Krylov.saunders_simon_yip

Krylov.saunders_simon_yip

Krylov.solution

Krylov.solution_count

Krylov.statistics

Krylov.sym_givens

Krylov.symmlq

Krylov.symmlq!

Krylov.to_boundary

Krylov.tricg

Krylov.tricg!

Krylov.trilqr

Krylov.trilqr!

Krylov.trimr

Krylov.trimr!

Krylov.usymlq

Krylov.usymlq!

Krylov.usymlqr

Krylov.usymlqr!

Krylov.usymqr

Krylov.usymqr!

Krylov.vec2str

Krylov.vector_to_matrix

Krylov.warm_start!

Internals

Krylov.FloatOrComplex — Type

FloatOrComplex{T}

Union type of T and Complex{T} where T is an AbstractFloat.

Krylov.warm_start! — Function

warmstart!(workspace, x0) warmstart!(workspace, x0, y0)

Warm-start the Krylov method associated with workspace using a user-provided initial guess, instead of starting from zero.

Krylov.kstdout — Constant

Default I/O stream for all Krylov methods.

Krylov.extract_parameters — Function

arguments = extract_parameters(ex::Expr)

Extract the arguments of an expression that is keyword parameter tuple. Implementation suggested by Mitchell J. O'Sullivan (@mosullivan93).

Base.show — Function

show(io, workspace; show_stats=true)

Statistics of workspace are displayed if show_stats is set to true.